+++ test/_env/util/auth.js
@@ -2,10 +2,12 @@
 
 const alice = `Basic ${Buffer.from("alice:alice").toString("base64")}`;
 const bob = `Basic ${Buffer.from("bob:bob").toString("base64")}`;
+const carol = `Basic ${Buffer.from("carol:carol").toString("base64")}`;
 const invalid = `Basic ${Buffer.from("invalid:invalid").toString("base64")}`;
 
 module.exports = {
   alice,
   bob,
+  carol,
   invalid,
 };